+++ Cmdlets/Private/ConvertTo-TeamViewerRoleAssignedUserGroup.ps1
@@ -13,6 +13,6 @@ function ConvertTo-TeamViewerRoleAssignedUserGroup {
         $result = New-Object -TypeName PSObject -Property $properties
         $result.PSObject.TypeNames.Insert(0, 'TeamViewerPS.RoleAssignedUserGroup')
 
-        $result
+        Write-Output $result
     }
 }