+++ Cmdlets/Private/Resolve-TeamViewerRoleId.ps1
@@ -8,10 +8,10 @@ function Resolve-TeamViewerRoleId {
 
     process {
         if ($Role.PSObject.TypeNames -contains 'TeamViewerPS.Role') {
-            [string]$Role.RoleID
+            Write-Output ([string]$Role.RoleID)
         }
         elseif ($Role -match '^[a-f0-9]{8}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{12}$') {
-            [string]$Role
+            Write-Output ([string]$Role)
         }
         else {
             throw "Invalid role identifier '$Role'. Must be either a [TeamViewerPS.Role] or [UUID] "